chore(penggajian): change data type from decimal to double

+++ database/migrations/2025_08_10_214136_add_kebijakan_fields_to_perusahaan.php
@@ -13,12 +13,12 @@ return new class extends Migration
     {
         Schema::table('perusahaan', function (Blueprint $table) {
             // Kebijakan Keterlambatan (sesuai aturan baru)
-            $table->decimal('potongan_keterlambatan', 15, 2)->default(50000)->after('jam_pulang');
+            $table->double('potongan_keterlambatan')->default(50000)->after('jam_pulang');
 
             // Konfigurasi Potongan BPJS Karyawan
-            $table->decimal('persen_bpjs_kesehatan', 5, 4)->default(0.0100)->comment('Default 1%')->after('potongan_keterlambatan');
-            $table->decimal('persen_bpjs_jht', 5, 4)->default(0.0200)->comment('Default 2%')->after('persen_bpjs_kesehatan');
-            $table->decimal('persen_bpjs_jp', 5, 4)->default(0.0100)->comment('Default 1%')->after('persen_bpjs_jht');
+            $table->double('persen_bpjs_kesehatan')->default(0.0100)->comment('Default 1%')->after('potongan_keterlambatan');
+            $table->double('persen_bpjs_jht')->default(0.0200)->comment('Default 2%')->after('persen_bpjs_kesehatan');
+            $table->double('persen_bpjs_jp')->default(0.0100)->comment('Default 1%')->after('persen_bpjs_jht');
             $table->bigInteger('batas_gaji_bpjs_kesehatan')->default(12000000)->after('persen_bpjs_jp');
             $table->bigInteger('batas_gaji_bpjs_pensiun')->default(10547400)->after('batas_gaji_bpjs_kesehatan');
         });
