+++ database/migrations/2025_08_10_214243_create_karyawan_table.php
@@ -35,7 +35,7 @@ return new class extends Migration {
             ])->nullable();
             $table->date('tanggal_mulai_bekerja')->nullable();
             $table->integer('kuota_cuti_tahunan')->default(12);
-            $table->decimal('gaji_pokok', 15, 2)->nullable();
+            $table->double('gaji_pokok')->nullable();
             $table->string('nomor_rekening', 50)->nullable();
             $table->string('nama_pemilik_rekening', 255)->nullable();
             $table->string('nomor_bpjs_kesehatan', 50)->nullable();
